+++ frontend/src/services/share.service.ts
@@ -156,7 +156,6 @@ const completeFile = async (
 
 const createReverseShare = async (
   shareExpiration: string,
-  maxShareSize: number,
   maxUseCount: number,
   sendEmailNotification: boolean,
   simplified: boolean,
@@ -165,7 +164,6 @@ const createReverseShare = async (
   return (
     await api.post("reverseShares", {
       shareExpiration,
-      maxShareSize: maxShareSize.toString(),
       maxUseCount,
       sendEmailNotification,
       simplified,